jobs/payment-jobs/requirements.txt
-e git+https://github.com/bcgov/sbc-common-components.git@d9b4616826f8dce6ec16540f24a88bfb530971c2#egg=sbc_common_components&subdirectory=python
-e git+https://github.com/bcgov/sbc-pay.git@e5af1153525824faf01d0e25a82d248e7ad05f72#egg=pay_api&subdirectory=pay-api
Flask-Caching==1.10.1
Flask-Migrate==2.7.0
Flask-Moment==1.0.2
Flask-OpenTracing==1.1.0
Flask-SQLAlchemy==2.5.1
Flask-Script==2.0.6
Flask==1.1.4
Jinja2==2.11.3
Mako==1.1.6
MarkupSafe==2.0.1
PyNaCl==1.4.0
SQLAlchemy-Continuum==1.3.11
SQLAlchemy-Utils==0.37.9
SQLAlchemy==1.3.24
Werkzeug==1.0.1
alembic==1.7.5
aniso8601==9.0.1
asyncio-nats-client==0.11.5
asyncio-nats-streaming==0.4.0
attrs==21.2.0
bcrypt==3.2.0
blinker==1.4
cachelib==0.4.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
click==7.1.2
croniter==1.1.0
cryptography==36.0.1
dpath==2.0.5
ecdsa==0.17.0
flask-jwt-oidc==0.3.0
flask-marshmallow==0.11.0
flask-restx==0.5.1
gunicorn==20.1.0
idna==3.3
itsdangerous==1.1.0
jaeger-client==4.8.0
jsonschema==4.3.1
marshmallow-sqlalchemy==0.25.0
marshmallow==3.14.1
minio==7.1.2
opentracing==2.4.0
paramiko==2.8.1
protobuf==3.19.1
psycopg2-binary==2.9.2
pyasn1==0.4.8
pycparser==2.21
pyrsistent==0.18.0
pysftp==0.2.9
python-dateutil==2.8.2
python-dotenv==0.19.2
python-jose==3.3.0
pytz==2021.3
requests==2.26.0
rsa==4.8
sentry-sdk==1.5.1
six==1.16.0
threadloop==1.0.2
thrift==0.15.0
tornado==6.1
urllib3==1.26.7
